Fix regression in server_test
